SOARAPPS-6287 trap repo errors

+++ fn_github/fn_github/components/funct_github_delete_branch.py
@@ -29,15 +29,14 @@ class FunctionComponent(AppFunctionComponent):
 
         yield self.status_message(f"Starting App Function: '{FN_NAME}'")
 
-        gh = GitHubHelper(fn_inputs.github_owner, fn_inputs.github_repo, self.options)
-
-        validate_fields([{"name": "base_url", "placeholder": "<https://base-url>"}],
-            self.app_configs)
-
         validate_fields(["github_owner", "github_repo", "github_branch"], fn_inputs)
 
-        results, err_msg = gh.delete_branch(fn_inputs.github_branch)
+        gh = GitHubHelper(fn_inputs.github_owner, fn_inputs.github_repo, self.options)
+        if gh.repo_err:
+            yield FunctionResult(None, success=False, reason=gh.repo_err)
+        else:
+            results, err_msg = gh.delete_branch(fn_inputs.github_branch)
 
-        yield self.status_message(f"Finished running App Function: '{FN_NAME}'")
+            yield self.status_message(f"Finished running App Function: '{FN_NAME}'")
 
-        yield FunctionResult(results, success=not bool(err_msg), reason=err_msg)
+            yield FunctionResult(results, success=not bool(err_msg), reason=err_msg)
